Make TimerPublisher usable everywhere
