+++ Sources/OpenCombineFoundation/RunLoop+Scheduler.swift
@@ -6,7 +6,6 @@
 //
 
 import CoreFoundation
-import Dispatch
 import Foundation
 import OpenCombine
